some fixes and improvements
